work on im chat and api modules

+++ src/wbs/imchat/core/api/ImChatTestRequest.java
@@ -15,6 +15,6 @@ class ImChatTestRequest {
 	String hello;
 
 	@DataAttribute
-	Integer world;
+	Long world;
 
 }
